backend/requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
click==8.3.0
coverage==7.11.0
Deprecated==1.2.18
fastapi==0.119.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
limits==5.6.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
python-decouple==3.8
python-json-logger==4.0.0
secure==1.0.1
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
wrapt==1.17.3
